Cash header, record pair, account cards — and one control vocabulary
The record control was a `+` and a `−` bolted onto a text button: three things
pretending to be one. It is now a joined pair and nothing else — add on the
left, take away on the right (the order a calculator and a number line already
taught) — with each half carrying its direction's colour AT REST, because a
bare glyph says nothing about money and meaning cannot live in hover.

Account-card actions became a full-bleed footer, flush to the card's edges and
split down the middle. Insetting them left a moat of dead padding on three
sides and made two buttons float inside a box instead of finishing it.

Month, record and accounts now share one header row; the month label was sized
for "September" and spent most of the year housing three letters, so it is
abbreviated and narrower, and managing accounts is a quiet icon beside the loud
actions rather than a worded button competing with them.

Control Center → Notifications: the mirrorchain master toggle was a bespoke
sliding pill while every per-type cell in the same table was a checkbox — two
vocabularies for one question. It is a master over a column, which is a real
difference, but that is exactly what `indeterminate` is for. Tri-state
preserved; only the shape changed.

+++ apps/web/src/user/portfolio/cashflow/MonthPicker.tsx
@@ -10,6 +10,9 @@ import { useI18n, useT } from '../../../i18n';
  * to say "July"; and moving one month — overwhelmingly the thing people do
  * here — meant editing a number rather than pressing a direction.
  *
+ * The month is abbreviated ("Jul 2026"): the control was sized for
+ * "September" and spent most of the year housing three letters.
+ *
  * Forward is disabled at the current month: the ledger cannot contain the
  * future, so an enabled arrow would only ever lead to an empty page.
  */
@@ -38,7 +41,7 @@ export function MonthPicker({ value, onChange }: MonthPickerProps) {
 
   const [year, index] = value.split('-').map(Number);
   const label = new Date(Date.UTC(year!, index! - 1, 1)).toLocaleDateString(locale, {
-    month: 'long',
+    month: 'short',
     timeZone: 'UTC',
     year: 'numeric',
   });
